fredtool*: split Flipnote steps and b9stool steps, move things around to align
This commit splits the rather large Flipnote exploit section into the
following includes:
    - exploit-flipnote
    - install-boot9strap-b9stool

This also rewords portions of the Fredtool steps in individual pages so
that the steps to launch Flipnote Studio are now in the Flipnote
Exploit section, so as to make the b9stool section flow through easily.

This also allows to remove more arguments within the `include_relative`
directive.