seedmiiner: don't use backplate year as year of manufacture (#2344)
After having this page in production for some time, it is clear that 
the year numbers simply do not matter. For instance, the original 2DS 
*also* shows 2011 on the backplate, yet it was only released in 2013.

Instead, add estimated manufacture dates for each model. These years 
and guesses are *not* proven; however, it gets us relatively close to 
what is expected based on the release date and how long it was being 
manufactured.